pyproject.toml
[project]
name = "deye-inverter-mqtt"
version = "2026.01.1"
description = "Reads Deye solar inverter metrics using Modbus over TCP and publishes them over MQTT."
authors = [
    { name = "Krzysztof Białek", email = "kbialek@users.noreply.github.com" }
]
license = { text = "Apache License 2" }
readme = "README.md"
requires-python = ">=3.13.0"

dependencies = [
    "libscrc==1.8.1",
    "paho-mqtt==1.6.1",
]

[dependency-groups]
dev = [
    "pytest>=7.2.2",
    "pytest-cov>=4.1.0",
    "pytest-mock",
    "black[d]>=23.1.0",
    "flake8>=6.0.0",
    "Flake8-pyproject>=1.2.3",
    "pre-commit>=3.2.0",
]

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
ignore = ["E203", "W503"]

[tool.coverage.run]
branch = true
relative_files = true

[tool.pytest.ini_options]
pythonpath = "src"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]